requirements.txt
fastapi==0.112.0
fastapi-cli==0.0.4
starlette==0.37.2
typer==0.12.3
passlib==1.7.4
cryptography==43.0.0
uvicorn==0.30.5
PyMySQL==1.1.1
SQLAlchemy==2.0.31
python-jose==3.3.0
boto3==1.34.152
python-dotenv==1.0.1
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.8.2
email_validator==2.2.0